feat(csg): Update sample utility funcs
