+++ src/bundles/csg/samples/primitives.js
@@ -15,14 +15,6 @@ let primitives = [
     torus
 ];
 
-function translate_x(entity, factor) {
-    return translate(entity, factor, 0, 0);
-}
-
-function translate_y(entity, factor) {
-    return translate(entity, 0, factor, 0);
-}
-
 let l = build_list(
     i => translate_y(
         translate_x(
