+++ src/bundles/csg/samples/ship.js
@@ -3,86 +3,7 @@
 // Mum: No, we have Source Academy ship at home
 // Source Academy ship at home:
 
-/* [Convenience Functions] */
-function translate_x(entity, factor) {
-    return translate(entity, factor, 0, 0);
-}
-
-function translate_y(entity, factor) {
-    return translate(entity, 0, factor, 0);
-}
-
-function translate_z(entity, factor) {
-    return translate(entity, 0, 0, factor);
-}
-
-function rotate_x(entity, factor) {
-    return rotate(entity, factor, 0, 0);
-}
-
-function rotate_y(entity, factor) {
-    return rotate(entity, 0, factor, 0);
-}
-
-function rotate_z(entity, factor) {
-    return rotate(entity, 0, 0, factor);
-}
-
-function scale_all(entity, factor) {
-    return scale(entity, factor, factor, factor);
-}
-
-function scale_x(entity, factor) {
-    return scale(entity, factor, 1, 1);
-}
-
-function scale_y(entity, factor) {
-    return scale(entity, 1, factor, 1);
-}
-
-function scale_z(entity, factor) {
-    return scale(entity, 1, 1, factor);
-}
-
 /* [Utility Functions] */
-function _get_shape_middle(shape, axis) {
-    let get = bounding_box(shape);
-    let start = get(axis, 'min');
-    let end = get(axis, 'max');
-    let length = end - start;
-    return start + (length / 2);
-}
-
-function _centre_at(shape, x, y, z) {
-    function calculate_offset(axis, centre_coord) {
-        return -_get_shape_middle(shape, axis) + centre_coord;
-    }
-
-    return translate(
-        shape,
-        calculate_offset('x', x),
-        calculate_offset('y', y),
-        calculate_offset('z', z)
-    );
-}
-
-function centre(shape) {
-    return _centre_at(shape, 0.5, 0.5, 0.5);
-}
-
-function centre_using(target, reference) {
-    return _centre_at(
-        target,
-        _get_shape_middle(reference, 'x'),
-        _get_shape_middle(reference, 'y'),
-        _get_shape_middle(reference, 'z')
-    );
-}
-
-function degrees_to_radians(degrees) {
-    return (degrees / 360) * (2 * math_PI);
-}
-
 function debug(shape) {
     let get = bounding_box(shape);
     let xStart = get('x', 'min');
