+++ src/bundles/csg/samples/steinmetz.js
@@ -1,26 +1,6 @@
 // Source §3
 // Cylinder intersects - Steinmetz solid
 
-function rotate_x(entity, factor) {
-    return rotate(entity, factor, 0, 0);
-}
-
-function rotate_y(entity, factor) {
-    return rotate(entity, 0, factor, 0);
-}
-
-function translate_y(entity, factor) {
-    return translate(entity, 0, factor, 0);
-}
-
-function translate_z(entity, factor) {
-    return translate(entity, 0, 0, factor);
-}
-
-function scale_all(entity, factor) {
-    return scale(entity, factor, factor, factor);
-}
-
 const A_ = cylinder(blue);
 const B_ = translate_y(rotate_x(cylinder(green), math_PI/2),1);
 const C_ = translate_z(rotate_y(cylinder(yellow),math_PI/2),1);
